Adjusted bomb strength in both modes to reflect the new hamster power!... err.. masS!
git-svn-id: svn://10.0.50.55/puffbomb/trunk@1928 d14347b8-48fd-0310-9e84-97907d434d85

+++ Code/Game/PuffBOMB/GolfGame/GolfGameEngine.cpp
@@ -173,7 +173,7 @@ void cGolfGameEngine::TurnBasedPlay() {
 						Engine2D::cImpulse(
 							BombPos,
 							// Inner Radius, Intensity, Tangent //
-							Real( 0 ), Real( 50 ), Real( 0 ),
+							Real( 0 ), Real( 30 ), Real( 0 ),
 							// Outer Radius, Intensity, Tangent //
 							Real( 512 ), Real( 0 ), Real( 0 )
 							)
