+++ Code/Game/PuffBOMB/Objects/BasicBomb/BasicBomb.cpp
@@ -44,7 +44,7 @@ bool cBasicBomb::Work() {
 			Engine2D::cImpulse(
 				Pos,
 				// Inner Radius, Intensity, Tangent //
-				Real( 0 ), Real( 50 ), Real( 0 ),
+				Real( 0 ), Real( 30 ), Real( 0 ),
 				// Outer Radius, Intensity, Tangent //
 				Real( 512 ), Real( 0 ), Real( 0 )
 				)
